+++ waf-openapi-20211001/alibabacloud_waf_openapi20211001/client.py
@@ -18791,6 +18791,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.cluster_id):
             query['ClusterId'] = request.cluster_id
+        if not UtilClient.is_unset(request.cluster_rule_resource_id):
+            query['ClusterRuleResourceId'] = request.cluster_rule_resource_id
         if not UtilClient.is_unset(request.instance_id):
             query['InstanceId'] = request.instance_id
         if not UtilClient.is_unset(request.region_id):
@@ -18838,6 +18840,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.cluster_id):
             query['ClusterId'] = request.cluster_id
+        if not UtilClient.is_unset(request.cluster_rule_resource_id):
+            query['ClusterRuleResourceId'] = request.cluster_rule_resource_id
         if not UtilClient.is_unset(request.instance_id):
             query['InstanceId'] = request.instance_id
         if not UtilClient.is_unset(request.region_id):